Move variables logic into its own file
