+++ src/config/config.go
@@ -8,9 +8,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/AlecAivazis/survey/v2"
 	"github.com/defenseunicorns/zarf/src/types"
-	"github.com/pterm/pterm"
 
 	"github.com/defenseunicorns/zarf/src/internal/message"
 	"github.com/defenseunicorns/zarf/src/internal/utils"
@@ -206,94 +204,8 @@ func LoadConfig(path string, filterByOS bool) error {
 	return nil
 }
 
-// FillActiveTemplate handles setting the active variables and reloading the base template.
-func FillActiveTemplate() error {
-	packageVariables, err := utils.FindYamlTemplates(&active, "###ZARF_PKG_VAR_", "###")
-	if err != nil {
-		return err
-	}
-
-	for key := range CommonOptions.SetVariables {
-		value := CommonOptions.SetVariables[key]
-		packageVariables[key] = &value
-	}
-
-	for key, value := range packageVariables {
-		if value == nil {
-			if !CommonOptions.Confirm {
-				if setVal, err := promptVariable(key, nil); err == nil {
-					packageVariables[key] = &setVal
-				} else {
-					return err
-				}
-			} else {
-				return fmt.Errorf("variable '%s' must be '--set' when using the '--confirm' flag", key)
-			}
-		}
-	}
-
-	templateMap := map[string]string{}
-	for key, value := range packageVariables {
-		// Variable keys are always uppercase in the format ###ZARF_PKG_VAR_KEY###
-		templateMap[strings.ToUpper(fmt.Sprintf("###ZARF_PKG_VAR_%s###", key))] = *value
-	}
-
-	return utils.ReloadYamlTemplate(&active, templateMap)
-}
-
-// SetActiveVariables handles setting the active variables used to template component files.
-func SetActiveVariables() error {
-	SetVariableMap = CommonOptions.SetVariables
-
-	for _, variable := range active.Variables {
-		if _, present := SetVariableMap[variable.Name]; !present {
-			if variable.Prompt && !CommonOptions.Confirm {
-				if val, err := promptVariable(variable.Name, variable.Default); err == nil {
-					SetVariableMap[variable.Name] = val
-				} else {
-					return err
-				}
-			} else if variable.Default != nil {
-				SetVariableMap[variable.Name] = *variable.Default
-			} else if variable.Prompt {
-				return fmt.Errorf("variable '%s' must be '--set' when using the '--confirm' flag", variable.Name)
-			}
-		}
-	}
-
-	return nil
-}
-
 func GetActiveConfig() types.ZarfPackage {
 	return active
-}
-
-// InjectImportedVariable determines if an imported package variable exists in the active config and adds it if not.
-func InjectImportedVariable(importedVariable types.ZarfPackageVariable) {
-	presentInActive := false
-	for _, configVariable := range active.Variables {
-		if configVariable.Name == importedVariable.Name {
-			presentInActive = true
-		}
-	}
-
-	if !presentInActive {
-		active.Variables = append(active.Variables, importedVariable)
-	}
-}
-
-// InjectImportedConstant determines if an imported package constant exists in the active config and adds it if not.
-func InjectImportedConstant(importedConstant types.ZarfPackageConstant) {
-	presentInActive := false
-	for _, configVariable := range active.Constants {
-		if configVariable.Name == importedConstant.Name {
-			presentInActive = true
-		}
-	}
-
-	if !presentInActive {
-		active.Constants = append(active.Constants, importedConstant)
-	}
 }
 
 // BuildConfig adds build information and writes the config to the given path
@@ -366,24 +278,4 @@ func isCompatibleComponent(component types.ZarfComponent, filterByOS bool) bool 
 	}
 
 	return validArch && validOS
-}
-
-func promptVariable(varName string, varDefault *string) (string, error) {
-	var value string
-
-	pterm.Println()
-
-	prompt := &survey.Input{
-		Message: "Please provide a value for '" + varName + "'",
-	}
-
-	if varDefault != nil {
-		prompt.Default = *varDefault
-	}
-
-	if err := survey.AskOne(prompt, &value); err != nil {
-		return "", err
-	}
-
-	return value, nil
 }
